ref(spans): Refactor span search filtering to mobx
