+++ static/app/components/events/interfaces/spans/traceView.tsx
@@ -1,172 +1,32 @@
 import {createRef, PureComponent} from 'react';
-import pick from 'lodash/pick';
+import {Observer} from 'mobx-react';
 
 import EmptyStateWarning from 'app/components/emptyStateWarning';
 import {t} from 'app/locale';
 import {Organization} from 'app/types';
 import {EventTransaction} from 'app/types/event';
-import {createFuzzySearch} from 'app/utils/createFuzzySearch';
 
 import * as CursorGuideHandler from './cursorGuideHandler';
 import * as DividerHandlerManager from './dividerHandlerManager';
 import DragManager, {DragManagerChildrenProps} from './dragManager';
-import {ActiveOperationFilter} from './filter';
 import TraceViewHeader from './header';
 import * as ScrollbarManager from './scrollbarManager';
 import SpanTree from './spanTree';
-import {ParsedTraceType, RawSpanType} from './types';
-import {generateRootSpan, getSpanID, getTraceContext} from './utils';
-
-type IndexedFusedSpan = {
-  span: RawSpanType;
-  indexed: string[];
-  tagKeys: string[];
-  tagValues: string[];
-  dataKeys: string[];
-  dataValues: string[];
-};
-
-type FuseResult = {
-  item: IndexedFusedSpan;
-  score: number;
-};
-
-export type FilterSpans = {
-  results: FuseResult[];
-  spanIDs: Set<string>;
-};
+import {ParsedTraceType} from './types';
+import {getTraceContext} from './utils';
+import WaterfallModel from './waterfallModel';
 
 type Props = {
   organization: Organization;
   event: Readonly<EventTransaction>;
   parsedTrace: ParsedTraceType;
-  searchQuery: string | undefined;
-  operationNameFilters: ActiveOperationFilter;
+  waterfallModel: WaterfallModel;
 };
 
-type State = {
-  filterSpans: FilterSpans | undefined;
-};
-
-class TraceView extends PureComponent<Props, State> {
-  constructor(props: Props) {
-    super(props);
-
-    this.state = {
-      filterSpans: undefined,
-    };
-
-    this.filterOnSpans(props.searchQuery);
-  }
-
-  componentDidUpdate(prevProps) {
-    if (prevProps.searchQuery !== this.props.searchQuery) {
-      this.filterOnSpans(this.props.searchQuery);
-    }
-  }
-
+class TraceView extends PureComponent<Props> {
   traceViewRef = createRef<HTMLDivElement>();
   virtualScrollBarContainerRef = createRef<HTMLDivElement>();
   minimapInteractiveRef = createRef<HTMLDivElement>();
-
-  async filterOnSpans(searchQuery: string | undefined) {
-    if (!searchQuery) {
-      // reset
-      if (this.state.filterSpans !== undefined) {
-        this.setState({
-          filterSpans: undefined,
-        });
-      }
-      return;
-    }
-
-    const {parsedTrace} = this.props;
-
-    const {spans} = parsedTrace;
-
-    const transformed: IndexedFusedSpan[] = [generateRootSpan(parsedTrace), ...spans].map(
-      (span): IndexedFusedSpan => {
-        const indexed: string[] = [];
-
-        // basic properties
-
-        const pickedSpan = pick(span, [
-          // TODO: do we want this?
-          // 'trace_id',
-          'span_id',
-          'start_timestamp',
-          'timestamp',
-          'op',
-          'description',
-        ]);
-
-        const basicValues: string[] = Object.values(pickedSpan)
-          .filter(value => !!value)
-          .map(value => String(value));
-
-        indexed.push(...basicValues);
-
-        // tags
-
-        let tagKeys: string[] = [];
-        let tagValues: string[] = [];
-        const tags: {[tag_name: string]: string} | undefined = span?.tags;
-
-        if (tags) {
-          tagKeys = Object.keys(tags);
-          tagValues = Object.values(tags);
-        }
-
-        const data: {[data_name: string]: any} | undefined = span?.data ?? {};
-
-        let dataKeys: string[] = [];
-        let dataValues: string[] = [];
-        if (data) {
-          dataKeys = Object.keys(data);
-          dataValues = Object.values(data).map(
-            value => JSON.stringify(value, null, 4) || ''
-          );
-        }
-
-        return {
-          span,
-          indexed,
-          tagKeys,
-          tagValues,
-          dataKeys,
-          dataValues,
-        };
-      }
-    );
-
-    const fuse = await createFuzzySearch(transformed, {
-      keys: ['indexed', 'tagKeys', 'tagValues', 'dataKeys', 'dataValues'],
-      includeMatches: false,
-      threshold: 0.6,
-      location: 0,
-      distance: 100,
-      maxPatternLength: 32,
-    });
-
-    const results = fuse.search<FuseResult>(searchQuery);
-
-    const spanIDs: Set<string> = results.reduce((setOfSpanIDs: Set<string>, result) => {
-      const spanID = getSpanID(result.item.span);
-
-      if (spanID) {
-        setOfSpanIDs.add(spanID);
-      }
-
-      return setOfSpanIDs;
-    }, new Set<string>());
-
-    this.setState({
-      filterSpans: {
-        results,
-        spanIDs,
-      },
-    });
-  }
 
   renderHeader = (dragProps: DragManagerChildrenProps, parsedTrace: ParsedTraceType) => (
     <TraceViewHeader
@@ -176,7 +36,7 @@ class TraceView extends PureComponent<Props, State> {
       trace={parsedTrace}
       event={this.props.event}
       virtualScrollBarContainerRef={this.virtualScrollBarContainerRef}
-      operationNameFilters={this.props.operationNameFilters}
+      operationNameFilters={this.props.waterfallModel.operationNameFilters}
     />
   );
 
@@ -191,7 +51,7 @@ class TraceView extends PureComponent<Props, State> {
       );
     }
 
-    const {organization, operationNameFilters} = this.props;
+    const {organization, waterfallModel} = this.props;
 
     return (
       <DragManager interactiveLayerRef={this.minimapInteractiveRef}>
@@ -211,15 +71,19 @@ class TraceView extends PureComponent<Props, State> {
                       dragProps={dragProps}
                     >
                       {this.renderHeader(dragProps, parsedTrace)}
-                      <SpanTree
-                        traceViewRef={this.traceViewRef}
-                        event={event}
-                        trace={parsedTrace}
-                        dragProps={dragProps}
-                        filterSpans={this.state.filterSpans}
-                        organization={organization}
-                        operationNameFilters={operationNameFilters}
-                      />
+                      <Observer>
+                        {() => (
+                          <SpanTree
+                            traceViewRef={this.traceViewRef}
+                            event={event}
+                            trace={parsedTrace}
+                            dragProps={dragProps}
+                            filterSpans={waterfallModel.filterSpans}
+                            organization={organization}
+                            operationNameFilters={waterfallModel.operationNameFilters}
+                          />
+                        )}
+                      </Observer>
                     </ScrollbarManager.Provider>
                   );
                 }}
